net: lib: azure_fota: Send FOTA report if pendingFwVersion is not found
The pre-condition for sending a FOTA report is that `fwUpdateStatus`,
`currentFwVersion`, and `pendingFwVersion` is present in the received
FOTA firmware object when parsing a received device twin.
However, when `jobID` whas introduced to the library, `pendingFwVersion`
may not be reported for all FOTA updates depending on the library having
obtained a jobID or not.

This patch addresses this issue by setting the `report_neeed` flag in
case `pendingFwVersion` is not present in the parsed device twin. This
ensures that a FOTA report including the current firmware version will
be sent when encountering this corner-case.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>